requirements.txt
asgiref==3.8.1
asn1crypto==1.5.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
comtypes==1.3.1
contourpy==1.2.0
cycler==0.12.1
dj-rest-auth==7.0.1
Django==5.1.3
django-allauth==65.4.1
django-bootstrap4==23.4
django-chartjs==2.3.0
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
esprima==4.0.1
et-xmlfile==1.1.0
Flask==3.0.0
fonttools==4.47.0
geojson==3.1.0
highcharts-core==1.5.1
highcharts-gantt==1.5.0
highcharts-maps==1.5.1
highcharts-stock==1.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
npm==0.1.1
numpy==1.26.1
openpyxl==3.1.2
optional-django==0.1.0
packaging==23.2
pandas==2.1.4
pg8000==1.30.2
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
PyJWT==2.9.0
pymongo==4.10.1
pyparsing==3.1.1
pypiwin32==223
pyproject-toml==0.0.10
python-dateutil==2.8.2
python-http-client==3.3.7
pyttsx3==2.90
pytz==2023.3.post1
pywin32==306
redis==5.2.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
schedule==1.2.2
scramp==1.4.4
sendgrid==6.11.0
setuptools==68.2.2
shapely==2.0.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tabulate==0.9.0
toml==0.10.2
topojson==1.7
tzdata==2023.3
urllib3==2.0.7
uv==0.1.44
validator-collection==1.5.0
Werkzeug==3.0.1
wheel==0.41.2
